[2021/04/16 이재현] 체인코드 요약 쿼리 변경 (JPQL -> QueryDsl)

+++ src/main/java/com/brchain/core/chaincode/dto/CcSummaryDto.java
@@ -8,6 +8,6 @@ import lombok.Data;
 public class CcSummaryDto {
 
 	private String conName; // 컨테이너 이름
-	private int    ccCnt;   // 체인코드 개수
+	private Long    ccCnt;   // 체인코드 개수
 
 }
